backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi>=0.129.2",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.46",
    "uvicorn[standard]>=0.41.0",
]
